# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a awesome web server.",
        "title": "higo Example API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1/",
    "paths": {
        "/ping/": {
            "post": {
                "description": "检查进程是否存活",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "状态检查",
                "responses": {
                    "200": {}
                }
            }
        },
        "/user/login/": {
            "post": {
                "description": "用户登录接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "user_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/user/logout/": {
            "delete": {
                "description": "登出用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {}
                }
            }
        },
        "/user/me/": {
            "get": {
                "description": "获取用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户详情",
                "responses": {
                    "200": {}
                }
            }
        },
        "/user/register/": {
            "post": {
                "description": "注册用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "昵称",
                        "name": "nickname",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "用户名",
                        "name": "user_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "确认密码",
                        "name": "password_confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        }
    }
}